+++ spec/integrational/protractor-config/name.js
@@ -3,11 +3,12 @@ var env = require('../environment');
 exports.config = {
   seleniumAddress: env.seleniumAddress,
   framework: 'jasmine2',
-  specs: ['../protractor/fail-test.js'],
+  specs: ['../protractor/angularjs-homepage-test.js'],
   plugins: [{
     path: '../../../index.js',
-    screenshotPath: '.tmp/failures',
-    screenshotOnExpect: 'failure',
-    screenshotOnSpec: 'failure'
-  }]
+    screenshotPath: '.tmp/name'
+  }],
+  capabilities: {
+    'browserName': env.capabilities.browserName
+  }
 };