+++ spec/integrational/protractor-config/pause.js
@@ -3,11 +3,11 @@ var env = require('../environment');
 exports.config = {
   seleniumAddress: env.seleniumAddress,
   framework: 'jasmine2',
-  specs: ['../protractor/fail-test.js'],
+  specs: ['../protractor/angularjs-homepage-pause-on-failure-test.js'],
   plugins: [{
     path: '../../../index.js',
-    screenshotPath: '.tmp/failures',
-    screenshotOnExpect: 'failure',
-    screenshotOnSpec: 'failure'
+    screenshotPath: '.tmp/pause',
+    pauseOn: 'failure',
+    verbose: 'debug'
   }]
 };